Add files via upload
A program linker is a software utility that plays a crucial role in the process of converting source code written in programming languages (such as C, C++, or Fortran) into executable machine code that a computer can understand and run. The process of converting source code into an executable form is known as compilation.