Add files via upload
A process scheduler is a core component of an operating system responsible for determining the order in which processes (also known as tasks or threads) are executed on a computer's CPU (Central Processing Unit). Its primary objective is to manage the allocation of CPU time among multiple processes efficiently and fairly, while also optimizing system performance and responsiveness.