Add files via upload
An I/O scheduler, also known as disk scheduler, is a software component in an operating system responsible for managing and optimizing the order in which input/output (I/O) requests are serviced for storage devices like hard disk drives (HDDs) and solid-state drives (SSDs). The primary goal of an I/O scheduler is to improve the overall I/O performance of the system by reducing disk seek times, minimizing I/O latency, and ensuring fair access to the storage device for different processes.